requirements.txt
openai
crewai
langchain
python-dotenv
